+++ api/src/main/java/org/allaymc/api/blockentity/interfaces/BlockEntityComparator.java
@@ -1,0 +1,12 @@
+package org.allaymc.api.blockentity.interfaces;
+
+import org.allaymc.api.blockentity.BlockEntity;
+import org.allaymc.api.blockentity.component.BlockEntityComparatorBaseComponent;
+
+/**
+ * Block entity for comparator blocks.
+ *
+ * @author daoge_cmd
+ */
+public interface BlockEntityComparator extends BlockEntity, BlockEntityComparatorBaseComponent {
+}